% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clusterTrimmer}
\alias{clusterTrimmer}
\title{Trim signal cluster based on pairwise r2}
\usage{
clusterTrimmer(r2, r2_threshold)
}
\arguments{
\item{r2}{the matrix of r2 values}

\item{r2_threshold}{the threshold on r2}
}
\value{
a numeric vector (possibly length 0)
of the signal clusters that should be trimmed/removed
}
\description{
This function identifies the clusters to remove
such that all remaining clusters have pairwise
r2 below a given threshold. It removes in
reverse order (starting with the last cluster
represented in the r2 matrix), and stops once
the pairwise r2 are all below the threshold.
It is recommended to run this function after
\code{\link{extractForSlope}}.
}
