% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extractForSlope}
\alias{extractForSlope}
\title{Extract SNPs from colocalization for slope fitting}
\usage{
extractForSlope(
  res,
  niter = 0,
  plot = TRUE,
  label = "Effect size of",
  a = "eQTL",
  b = "GWAS"
)
}
\arguments{
\item{res}{list with the following named elements:
\itemize{
\item \code{beta_hat_a} - list of point estimates of coefficients for A from colocalization
\item \code{beta_hat_b} - " " for B
\item \code{sd_a} - list of sampling SD for \code{beta_hat_a} (in practice original
SE are provided here)
\item \code{sd_b} - " " for \code{beta_hat_b} " "
\item {alleles} (optional) list of data.frame with allele information
}}

\item{niter}{number of iterations of EM to run
for mclust, if set to 0, only the maximum
variant (in terms of A effect size) per
signal cluster is output. Default is to not
run clustering, but to take the SNP with the
largest effect size in A (in the positive direction)}

\item{plot}{logical, draw a before after of which
variants will be included for slope estimation}

\item{label}{what preceeds \code{a} and \code{b} in
the x- and y-axis labels}

\item{a}{name of A experiment}

\item{b}{name of B experiment}
}
\value{
list of vectors of the first four arguments,
collapsed now across signal clusters, representing
variants with positive effect on A. So the null variants
have been removed (and any variants per cluster that
indicated a negative effect on A). If \code{alleles}
data.frames were included in the input, they will
also be passed through as a single data.frame with the
selected SNPs per signal cluster
}
\description{
Extracts one or more SNPs from each signal cluster
based on the posterior estimate of the effect size
for A (largest effect size in the positive direction).
After running this function, it is recommended to use
\code{\link{clusterTrimmer}} to remove signal clusters
that are too highly correlated.
}
