% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitBetaEcaviar}
\alias{fitBetaEcaviar}
\title{Fit the beta eCAVIAR model}
\usage{
fitBetaEcaviar(data)
}
\arguments{
\item{data}{a list with the following named elements:
\itemize{
\item{nsnp - number of SNPs}
\item{beta_hat_a - vector of length nsnp, estimated coefficients for A}
\item{beta_hat_b - " " for B}
\item{se_a - vector of length nsnp, standard errors for beta_hat_a}
\item{se_b - " " for beta_hat_b}
\item{Sigma_a - correlation matrix of SNPs for A, dimension should be nsnp x nsnp}
\item{Sigma_b - " " for B (this could be different for different LD structures)}
}}
}
\description{
Fit the beta eCAVIAR model
}
