% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitBetaMixture}
\alias{fitBetaMixture}
\title{Fit the beta mixture model}
\usage{
fitBetaMixture(data)
}
\arguments{
\item{data}{a list with the following named elements:
\itemize{
\item{nsnp - vector, number of SNPs per cluster}
\item{beta_hat_a - vector of length sum(nsnp), first step point estimates of beta for A}
\item{beta_hat_b - " " for B}
\item{sd_a - vector of length sum(nsnp), first step posterior SD for beta for A}
\item{sd_b - " " for B}
}}
}
\description{
Fit the beta mixture model
}
