% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitMixture}
\alias{fitMixture}
\title{Fit the mixture model for null and large effects}
\usage{
fitMixture(
  beta_hat_a,
  beta_hat_b,
  sd_a,
  sd_b,
  sigma_0a,
  sigma_0b,
  sigma_1a,
  alpha_sd,
  mu_loc,
  mu_sd,
  sigma1b_sd,
  ...
)
}
\arguments{
\item{beta_hat_a}{vector of length sum(nsnp), first step point estimates of beta for A}

\item{beta_hat_b}{" " for B}

\item{sd_a}{vector of length sum(nsnp), first step posterior SD for beta for A}

\item{sd_b}{" " for B}

\item{sigma_0a}{prior SD of the null component for A}

\item{sigma_0b}{" " for B}

\item{sigma_1a}{prior SD of the non-null component for beta for A (SD for B is fitted parameter)}

\item{alpha_sd}{prior SD for alpha}

\item{mu_loc}{center of prior for mu}

\item{mu_sd}{prior SD for mu}

\item{sigma1b_sd}{prior SD for sigma_1b}

\item{...}{further arguments passed to rstan::sampling}
}
\description{
Fit the mixture model for null and large effects
}
