% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitSlope}
\alias{fitSlope}
\title{Fit the beta slope model: effect of A on B}
\usage{
fitSlope(
  res,
  sd_beta = NULL,
  mu_alpha = NULL,
  sd_alpha = NULL,
  sd_sigma = 1,
  ...
)
}
\arguments{
\item{res}{list with the following named elements:
\itemize{
\item beta_hat_a - vector of length sum(nsnp), first step point estimates of beta for A
\item beta_hat_b - " " for B
\item sd_a - vector of length sum(nsnp), first step posterior SD for beta for A
\item sd_b - " " for B
}}

\item{sd_beta}{prior SD for beta A}

\item{mu_alpha}{prior mean for alpha}

\item{sd_alpha}{prior SD for alpha}

\item{sd_sigma}{prior SD for sigma}

\item{...}{further arguments passed to rstan::sampling}
}
\description{
Fit the beta slope model: effect of A on B
}
