% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{flipAllelesAndGather}
\alias{flipAllelesAndGather}
\title{Flip alleles and gather results into lists}
\usage{
flipAllelesAndGather(sum_stat, ld_mat)
}
\arguments{
\item{sum_stat}{list of summary statistic tables}

\item{ld_mat}{list of LD matrices}
}
\value{
list with estimated coefficients, standard
errors, LD matrix, and allele table
}
\description{
Flip alleles and gather results into lists
}
