% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeSimDataForMrlocus}
\alias{makeSimDataForMrlocus}
\title{Make simple simulated summary data}
\usage{
makeSimDataForMrlocus(
  nsnp = c(7:10),
  idx = 5,
  alpha = 0.5,
  sigma = 0.05,
  betas = 1:4,
  se = 0.25,
  n_mult = 1
)
}
\arguments{
\item{nsnp}{number of SNPs per signal cluster}

\item{idx}{the causal SNP (same per cluster for
simplicity)}

\item{alpha}{the true slope of B coefficients over A coefficients}

\item{sigma}{the SD of true B coefficients around the conditional
values given true A coefficients}

\item{betas}{the true A coefficients}

\item{se}{the standard errors for betas}

\item{n_mult}{how many more samples the B study has}
}
\value{
a list of \code{beta_hat_a}, \code{beta_hat_b},
\code{se_a}, \code{se_b},
\code{Sigma_a}, \code{Sigma_b} (themselves lists), and \code{alleles}
(a list of data.frames each with
\code{id}, \code{ref}, \code{eff} for the SNP id,
reference allele, and effect allele).
}
\description{
Make simple simulated summary data
}
