% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus-package.R
\docType{package}
\name{mrlocus-package}
\alias{mrlocus-package}
\alias{mrlocus}
\title{The 'mrlocus' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
