% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{normalizedAllelicSpread}
\alias{normalizedAllelicSpread}
\title{Estimate the normalized allelic spread}
\usage{
normalizedAllelicSpread(fit)
}
\arguments{
\item{fit}{a list output by \code{estimateSlope}}
}
\value{
the normalized allelic spread, a numeric
}
\description{
A useful statistic is to normalize the estimate of
sigma from \code{estimateSlope}, the dispersion,
with respect to typical mediated effect sizes.
This helps in particular to compare sigma across
different GWAS traits, which may have different
scale.
}
\details{
We define the normalized allelic spread as an estimate of sigma
divided by something called the "mean mediated effect".
The mean mediated effect is defined as the mean
QTL effect size multiplied by the estimate of the slope.
So in the \code{plotMrlocus} plot, the mean mediated
effect is found by taking the mean among the
spread of points on the x-axis, then going up to the
linear trend. This function outputs the normalized
allelic spread which is again: sigma / mean mediated
effect.
}
