% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotInitEstimates}
\alias{plotInitEstimates}
\title{Plot initial estimates over signal clusters}
\usage{
plotInitEstimates(x, label = "Effect size of", a = "eQTL", b = "GWAS")
}
\arguments{
\item{x}{list of signal clusters data with \code{beta_hat_a}
and \code{beta_hat_b} lists}

\item{label}{what preceeds \code{a} and \code{b} in
the x- and y-axis labels}

\item{a}{name of A experiment}

\item{b}{name of B experiment}
}
\description{
Plot initial estimates over signal clusters
}
