% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{collapseHighCorSNPs}
\alias{collapseHighCorSNPs}
\title{Collapse high correlation SNPs}
\usage{
collapseHighCorSNPs(sum_stat, ld_mat, thresh = 0.95, plot = TRUE)
}
\arguments{
\item{sum_stat}{list of summary statistic tables}

\item{ld_mat}{list of LD matrices}

\item{plot}{logical, draw a before/after grid of plots}

\item{threshold}{threshold on absolute value of
correlation for collapsing, e.g. will collapse
if SNPs are more correlated (or anti-correlated)
than this amount}
}
\value{
list with modified ld_mat and sum_stat lists
}
\description{
Collapse high correlation SNPs
}
