% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extractForSlope}
\alias{extractForSlope}
\title{Extract SNPs from colocalization for slope fitting}
\usage{
extractForSlope(res, niter = 0, plot = TRUE, a = "eQTL", b = "GWAS")
}
\arguments{
\item{res}{list with the following named elements:
\itemize{
\item beta_hat_a - vector of length sum(nsnp), first step point estimates of beta for A
\item beta_hat_b - " " for B
\item sd_a - vector of length sum(nsnp), first step posterior SD (or SE) for beta for A 
\item sd_b - " " for B
}}

\item{niter}{number of iterations of EM to run
for Mclust, if set to 0, only the maximum
variant (in terms of A effect size) per
signal cluster is output.}

\item{plot}{logical, draw a before after of which
variants will be included for slope estimation}

\item{a}{name of A experiment}

\item{b}{name of B experiment}
}
\value{
list of vectors of the first four arguments,
collapsed now across signal clusters, representing
variants with positive effect on A. So the null variants
have been removed (and any variants per cluster that
indicated a negative effect on A)
}
\description{
Extract SNPs from colocalization for slope fitting
}
