% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extractForSlope}
\alias{extractForSlope}
\title{Extract SNPs from colocalization for slope fitting}
\usage{
extractForSlope(beta_hat_a, beta_hat_b, sd_a, sd_b, niter = 3, plot = TRUE)
}
\arguments{
\item{beta_hat_a}{list of coefficients for A}

\item{beta_hat_b}{" " for B}

\item{sd_a}{list of posterior SD for A}

\item{sd_b}{" " for B}

\item{niter}{number of iterations of EM to run
for Mclust}
}
\value{
list of vectors of the first four arguments,
collapsed now across signal clusters, representing
variants with positive effect on A. So the null variants
have been removed (and any variants per cluster that
indicated a negative effect on A)
}
\description{
Extract SNPs from colocalization for slope fitting
}
