% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitBetaMixture}
\alias{fitBetaMixture}
\title{Fit the beta mixture model}
\usage{
fitBetaMixture(data)
}
\arguments{
\item{data}{a list with the following named elements:
\itemize{
\item{nsnp - maximum number of SNPs}
\item{ncond - number of conditional signal clusters}
\item{n - number of SNPs per conditonal signal cluster}
\item{beta_hat_a - matrix nsnp x ncond, first step point estimates of beta for A}
\item{beta_hat_b - " " for B}
\item{sd_a - matrix nsnp x ncond, first step posterior SD for beta for A}
\item{sd_b - " " for B}
}}
}
\description{
Fit the beta mixture model
}
