% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitBetaMixture}
\alias{fitBetaMixture}
\title{Fit the beta mixture model}
\usage{
fitBetaMixture(
  nsnp,
  beta_hat_a,
  beta_hat_b,
  sd_a,
  sd_b,
  sigma_0a = 0.5,
  sigma_0b = 0.5,
  sigma_1a = 2,
  alpha_sd = 1,
  mu_loc = 8,
  mu_sd = 2,
  sigma1b_sd = 3,
  ...
)
}
\arguments{
\item{nsnp}{vector, number of SNPs per signal cluster}

\item{beta_hat_a}{vector of length sum(nsnp), first step point estimates of beta for A}

\item{beta_hat_b}{" " for B}

\item{sd_a}{vector of length sum(nsnp), first step posterior SD for beta for A}

\item{sd_b}{" " for B}

\item{sigma_0a}{prior SD of the null component for A}

\item{sigma_0b}{" " for B}

\item{sigma_1a}{prior SD of the non-null component for beta for A (SD for B is fitted parameter)}

\item{alpha_sd}{prior SD for alpha}

\item{mu_loc}{center of prior for mu}

\item{mu_sd}{prior SD for mu}

\item{sigma1b_sd}{prior SD for sigma_1b}

\item{...}{further arguments passed to rstan::sampling}
}
\description{
Fit the beta mixture model
}
