% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlocus.R
\name{fitSlope}
\alias{fitSlope}
\title{Fit the beta slope model: effect of A on B}
\usage{
fitSlope(res, alpha_mu = NULL, alpha_sd = NULL, sigma_sd = 1, ...)
}
\arguments{
\item{res}{list with the following named elements:
\itemize{
\item beta_hat_a - vector of length sum(nsnp), first step point estimates of beta for A
\item beta_hat_b - " " for B
\item sd_a - vector of length sum(nsnp), first step posterior SD for beta for A
\item sd_b - " " for B
}}

\item{alpha_mu}{prior mean for alpha}

\item{alpha_sd}{prior SD for alpha}

\item{sigma_sd}{prior SD for sigma}

\item{...}{further arguments passed to rstan::sampling}
}
\description{
Fit the beta slope model: effect of A on B
}
