% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{flipAllelesAndGather}
\alias{flipAllelesAndGather}
\title{Flip alleles and gather results into lists}
\usage{
flipAllelesAndGather(
  sum_stat,
  ld_mat,
  ld_mat2 = NULL,
  a,
  b,
  ref,
  eff,
  beta,
  se,
  a2_plink,
  a2_plink_mat2 = NULL,
  snp_id,
  sep,
  ab_last = TRUE,
  alleles_same = FALSE,
  plot = TRUE
)
}
\arguments{
\item{sum_stat}{list of summary statistic tables}

\item{ld_mat}{list of LD matrices}

\item{ld_mat2}{optional second list of LD matrices
(for different populations). it will be returned
alongside the first \code{ld_mat}, which is used
for the allele flipping. the second list of LD
matrices is just flipped in the same way}

\item{a}{name of A in columns of sum_stat ("eQTL")}

\item{b}{name of B ("GWAS")}

\item{ref}{name of reference allele}

\item{eff}{name of effect allele}

\item{beta}{name of estimated coefficient}

\item{se}{name of standard error}

\item{a2_plink}{name of the column representing the a2 allele
(reference allele) according to plink. the default for plink v1.9
and earlier was to reset a2 to the major allele, unless
an optional flag was used. in plink v2.0 and onward, one
should check to see which allele is used as reference for
calculating the LD matrix}

\item{a2_plink_mat2}{name of the column representing
the a2 allele for the second LD matrix, \code{ld_mat2}
(needed only if \code{ld_mat2} was specified)}

\item{snp_id}{name of SNP id}

\item{sep}{character separator in column names that involve A/B}

\item{ab_last}{logical, A/B descriptor is last in column names
(e.g. "beta_eqtl", "se_eqtl"))}

\item{alleles_same}{logical, A/B/LD matrix alleles are identical}

\item{plot}{logical, draw a scatterplot of the flipped betas}
}
\value{
list with estimated coefficients, standard
errors, LD matrix, and allele table
}
\description{
Flip alleles and gather results into lists
}
