% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{flipAllelesAndGather}
\alias{flipAllelesAndGather}
\title{Flip alleles and gather results into lists}
\usage{
flipAllelesAndGather(
  sum_stat,
  ld_mat,
  a,
  b,
  ref,
  eff,
  beta,
  se,
  a2_plink,
  snp_id,
  sep,
  ab_last = TRUE,
  plot = TRUE
)
}
\arguments{
\item{sum_stat}{list of summary statistic tables}

\item{ld_mat}{list of LD matrices}

\item{a}{name of A in columns of sum_stat ("eQTL")}

\item{b}{name of B ("GWAS")}

\item{ref}{name of reference allele}

\item{eff}{name of effect allele}

\item{beta}{name of estimated coefficient}

\item{se}{name of standard error}

\item{a2_plink}{name of the column representing the a2 allele
(reference allele) according to plink. the default for plink v1.9
and earlier was to reset a2 to the major allele, unless
an optional flag was used. in plink v2.0 and onward, one
should check to see which allele is used as reference for
calculating the LD matrix}

\item{snp_id}{name of SNP id}

\item{sep}{character separator in column names that involve A/B}

\item{ab_last}{A/B descriptor is last in column names
(e.g. "beta_eqtl", "se_eqtl"))}

\item{plot}{logical, draw a scatterplot of the flipped betas}
}
\value{
list with estimated coefficients, standard
errors, LD matrix, and allele table
}
\description{
Flip alleles and gather results into lists
}
