% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeSimDataForMrlocus}
\alias{makeSimDataForMrlocus}
\title{Make simulated data for mrlocus}
\usage{
makeSimDataForMrlocus(
  nsnp = c(7:10),
  idx = 5,
  alpha = 0.5,
  sigma = 0.05,
  betas = 1:4,
  se = 0.25,
  n_mult = 1
)
}
\arguments{
\item{nsnp}{number of SNPs per signal cluster}

\item{idx}{the causal SNP (same per cluster for
simplicity)}

\item{alpha}{the true slope of B coefficients over A coefficients}

\item{sigma}{the SD of true B coefficients around the conditional
values given true A coefficients}

\item{betas}{the true A coefficients}

\item{se}{the standard errors for betas}

\item{n_mult}{how many more samples the B study has}
}
\value{
a list of beta_hat_a, beta_hat_b, se_a, and se_b,
Sigma_a, and Sigma_b (themselves lists)
}
\description{
Make simulated data for mrlocus
}
