% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotInitEstimates}
\alias{plotInitEstimates}
\title{Plot initial estimates over signal clusters}
\usage{
plotInitEstimates(x, a = "eQTL", b = "GWAS")
}
\arguments{
\item{x}{list of signal clusters data with beta_hat_a
and beta_hat_b lists}

\item{a}{name of A experiment}

\item{b}{name of B experiment}
}
\description{
Plot initial estimates over signal clusters
}
