% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotMrlocus}
\alias{plotMrlocus}
\title{Plot mrlocus estimates}
\usage{
plotMrlocus(res, a = "eQTL", b = "GWAS", ...)
}
\arguments{
\item{res}{the output from fitSlope}

\item{a}{name of A experiment}

\item{b}{name of B experiment}

\item{...}{arguments passed to plot()}
}
\description{
Plot mrlocus estimates
}
